+++ test/api.spec.js
@@ -16,9 +16,9 @@ describe('api', () => {
         helper = require('../static/js/helpers')
     })
 
-    afterEach(() => {
-        fetch.resetMocks();
-    })
+    // afterEach(() => {
+    //     fetch.resetMocks();
+    // })
 
     describe('requests', () => {
         describe('get requests ', () => {
@@ -75,18 +75,40 @@ describe('api', () => {
                 expect(fetch).toHaveBeenCalled();
             })
 
-            test.skip('it makes a request to the giphy Api', async () => {
+            // test.skip('it makes a request to the giphy Api', async () => {
 
                 
-                const fakeEvent = {
-                    preventDefault: jest.fn(),
-                    target: {
-                        searchTerm: {value: 'hello'}
-                    }
-                }
-                const hello = await api.getGiphs(fakeEvent)
-                expect(fetch).toHaveBeenCalled();
-            })
+            //     const fakeEvent = {
+            //         preventDefault: jest.fn(),
+            //         target: {
+            //             searchTerm: {value: 'hello'}
+            //         }
+            //     }
+            //     const hello = await api.getGiphs(fakeEvent)
+            //     expect(fetch).toHaveBeenCalled();
+            // })
+
+            global.fetch = jest.fn(() =>
+                Promise.resolve({
+                json: () => Promise.resolve( (res) => gifs = res.body  ),
+                })
+            );
+  
+
+            test('it makes a request to the giphy Api', async () => {
+
+
+            const fakeEvent = {
+            preventDefault: jest.fn(),
+            target: {
+             searchTerm: {value: 'hello'}
+             }
+            }
+    
+            const giphysent = await getGiphs(fakeEvent);
+
+            expect(fetch).toHaveBeenCalledTimes(1);
+            });
         })
 
     })
